+++ src/Components/StaticsTab/StaticsList/ContentBar/Content/AverageScore.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function AverageScore(props) {
+    return (
+        <td>
+            <div className={props.className}>
+                {2.5 + ":1"}
+            </div>
+        </td>
+    )
+}
