+++ src/Components/StaticsTab/StaticsList/ContentBar/Content/Champions.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+export default function Champions(props) {
+    return (
+        <td>
+            <div className={props.className}>
+                <img src={""} alt={""}/>
+                <div className="champion-name">
+                    애쉬
+                </div>
+            </div>
+        </td>
+    )
+}
