+++ src/Components/StaticsTab/StaticsList/ContentBar/Content/PlayCount.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function PlayCount(props) {
+    return (
+        <td>
+            <div className={props.className}>
+                {123123}
+            </div>
+        </td>
+    )
+}
