+++ src/Components/StaticsTab/StaticsList/ContentBar/Content/Rank.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default function Rank(props) {
+    return (
+        <td>
+            <div className={props.className}>
+                {props.rank}
+            </div>
+        </td>
+    )
+}
