+++ src/Components/StaticsTab/StaticsList/ContentBar/Content/WinRate.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+export default function WinRate(props) {
+    return (
+        <td>
+            <div className={props.className}>
+                <div
+                    className="winRate-graph"
+                    style={{width: "53.62%"}}>
+                </div>
+                <div className="winRate-number">
+                    {53.62 + "%"}
+                </div>
+            </div>
+        </td>
+    )
+}
