+++ src/Components/StaticsTab/StaticsList/ContentBar/ContentBarContent.scss
@@ -1,16 +1,11 @@
 .statics-listContentBar-background {
-    display: flex;
     border-bottom: 1px solid #b8bfbf;
-    justify-content: space-between;
-    align-items: center;
     padding: 2px;
 
     .rank {
         display: flex;
         justify-content: center;
         align-items: center;
-        // border: 1px solid black;
-        width: 4%;
         padding: 5px 10px;
         color: #555e5e;
     }
@@ -19,8 +14,6 @@
         display: flex;
         justify-content: flex-start;
         align-items: center;
-        // border: 1px solid black;
-        width: 20%;
         padding: 5px 7px;
 
         img {
@@ -41,8 +34,6 @@
         display: flex;
         justify-content: flex-start;
         align-items: center;
-        // border: 1px solid black;
-        width: 20%;
         padding: 5px 7px;
 
         .winRate-graph {
@@ -62,8 +53,6 @@
         display: flex;
         justify-content: flex-start;
         align-items: center;
-        // border: 1px solid black;
-        width: 13%;
         padding: 5px 7px;
         color: #555e5e;
     }
@@ -72,8 +61,6 @@
         display: flex;
         justify-content: flex-start;
         align-items: center;
-        // border: 1px solid black;
-        width: 13%;
         padding: 5px 7px;
         font-weight: bold;
         color: #555e5e;
@@ -83,8 +70,6 @@
         display: flex;
         justify-content: flex-start;
         align-items: center;
-        // border: 1px solid black;
-        width: 15%;
         padding: 5px 7px;
 
         .cs-score-graph {
@@ -103,8 +88,7 @@
     .gold-score {
         display: flex;
         justify-content: flex-start;
-        // border: 1px solid black;
-        width: 15%;
+        align-items: center;
         padding: 5px 7px;
 
         .gold-score-graph {
