+++ src/Components/StaticsTab/StaticsList/TitleBar/Content/TitleBarContent.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+export default function StaticsListTitleBarContent(props) {
+    return (
+        <th>
+            <div className={props.className}>
+                <div>
+                    {props.title}
+                </div>
+                <div>
+                    {props.icon}
+                </div>
+            </div>         
+        </th>
+    )
+}
