+++ src/Components/StaticsTab/StaticsList/TitleBar/Content/TitleBarContent.scss
@@ -3,6 +3,4 @@
 //     border: 1px solid black;
 //     justify-content: center;
 //     align-items: center;
-
-    
 // }
