+++ src/Components/StaticsTab/StaticsList/TitleBar/Content/index.js
@@ -1,0 +1,1 @@
+export { default } from './TitleBarContent';
