+++ src/Components/StaticsTab/StaticsList/TitleBar/index.js
@@ -1,0 +1,1 @@
+export { default } from './TitleBar';
