--- src/Components/StaticsTab/StaticsListContentBar/Content/AverageScore.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function AverageScore(props) {
-    return (
-        <div className={props.className}>
-            {2.5 + ":1"}
-        </div>
-    )
-}
