--- src/Components/StaticsTab/StaticsListContentBar/Content/Champions.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-export default function Champions(props) {
-    return (
-        <div className={props.className}>
-            <img src={""} alt={""}/>
-            <div className="champion-name">
-                애쉬
-            </div>
-        </div>
-    )
-}
