--- src/Components/StaticsTab/StaticsListContentBar/Content/PlayCount.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function PlayCount(props) {
-    return (
-        <div className={props.className}>
-            {123123}
-        </div>
-    )
-}
