--- src/Components/StaticsTab/StaticsListContentBar/Content/Rank.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Rank(props) {
-    return (
-        <div className={props.className}>
-            {props.rank}
-        </div>
-    )
-}
