--- src/Components/StaticsTab/StaticsListTitleBar/StaticsListTitleBar.scss
@@ -1,71 +1,0 @@
-.statics-listTitleBar-background {
-    display: flex;
-    // border: 1px solid black;
-    justify-content: space-between;
-    align-items: flex-start;
-    padding: 5px;
-    border-bottom: 1px solid #b8bfbf;
-    color: #555e5e;
-
-    .rank {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 4%;
-        padding: 5px 10px;
-    }
-
-    .champion {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 20%;
-        padding: 5px 7px;
-    }
-
-    .win-rate {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 20%;
-        padding: 5px 7px;
-    }
-
-    .play-count {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 13%;
-        padding: 5px 7px;
-    }
-    
-    .average-score {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 13%;
-        padding: 5px 7px;
-    }
-
-    .cs-score {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        // border: 1px solid black;
-        width: 15%;
-        padding: 5px 7px;
-    }
-
-    .gold-score {
-        display: flex;
-        justify-content: space-between;
-        // border: 1px solid black;
-        width: 15%;
-        padding: 5px 7px;
-    }
-}
