--- src/Components/StaticsTab/StaticsListTitleBarContent/StaticsListTitleBarContent.js
@@ -1,14 +1,0 @@
-import React from 'react'
-
-export default function StaticsListTitleBarContent(props) {
-    return (
-        <div className={props.className}>
-            <div>
-                {props.title}
-            </div>
-            <div>
-                {props.icon}
-            </div>         
-        </div>
-    )
-}
