+++ src/Pages/StaticsChampions/staticsChampions.scss
@@ -14,10 +14,16 @@
         }
 
         .statics-champions-content {
+            border-collapse: collapse;
+            border-spacing: 0;
             margin-top: 20px;
             border: 1px solid #b8bfbf;
-            width: 100%;
-            //padding: 5px 0;
+            width: 101%;
+
+            tbody {
+                display: table-row-group;
+                vertical-align: middle;
+            }
         }
     }
 }
